[fix] No.107 Chapter19-1
* コード修正
